refactor: move eeprom transfer to uhk-usb package
